+++ alma/lib/Helpers/DateHelper.php
@@ -31,7 +31,6 @@ if (!defined('_PS_VERSION_')) {
 /**
  * Class DateHelper.
  *
- * @deprecated use DatesHelper instead
  *
  * Use for method date
  */
@@ -45,12 +44,12 @@ class DateHelper
      *
      * @return array|string[]
      */
-    public static function getDatesInInterval($from, $first)
+    public function getDatesInInterval($from, $first)
     {
         $to = strtotime('-1 day');
         $datesInInterval = [];
-        $startTimestamp = self::extractTimestampWithoutTime($from);
-        $firstWithoutTime = self::extractTimestampWithoutTime($first);
+        $startTimestamp = $this->extractTimestampWithoutTime($from);
+        $firstWithoutTime = $this->extractTimestampWithoutTime($first);
 
         for ($date = $startTimestamp; $date <= $to; $date = strtotime('+1 day', $date)) {
             if ($date >= $firstWithoutTime) {
@@ -61,7 +60,6 @@ class DateHelper
         return $datesInInterval;
     }
 
-
     /**
      * check if is the same date without time
      *
@@ -70,9 +68,9 @@ class DateHelper
      *
      * @return bool
      */
-    public static function isSameDay($today, $day)
+    public function isSameDay($today, $day)
     {
-        return self::extractDateWithoutTime($today) === self::extractDateWithoutTime($day);
+        return $this->extractDateWithoutTime($today) === $this->extractDateWithoutTime($day);
     }
 
     /**
@@ -82,9 +80,9 @@ class DateHelper
      *
      * @return int
      */
-    private static function extractTimestampWithoutTime($timestamp)
+    private function extractTimestampWithoutTime($timestamp)
     {
-        return strtotime(self::extractDateWithoutTime($timestamp));
+        return strtotime($this->extractDateWithoutTime($timestamp));
     }
 
     /**
@@ -94,10 +92,11 @@ class DateHelper
      *
      * @return string
      */
-    private static function extractDateWithoutTime($timestamp)
+    private function extractDateWithoutTime($timestamp)
     {
         return date('Y-m-d', $timestamp);
     }
+
 
     /**
      * format date by locale
@@ -107,7 +106,7 @@ class DateHelper
      *
      * @return string date
      */
-    public static function getDateFormat($locale, $timestamp)
+    public function getDateFormat($locale, $timestamp)
     {
         try {
             if (class_exists(\IntlDateFormatter::class)) {
@@ -119,7 +118,7 @@ class DateHelper
             // We don't need to deal with this Exception because a fallback exists in default return statement
         }
 
-        return static::getFrenchDateFormat($timestamp);
+        return $this->getFrenchDateFormat($timestamp);
     }
 
     /**
@@ -129,7 +128,7 @@ class DateHelper
      *
      * @return string
      */
-    protected static function getFrenchDateFormat($timestamp)
+    protected function getFrenchDateFormat($timestamp)
     {
         $date = new \DateTime();
         $date->setTimestamp($timestamp);
@@ -142,7 +141,7 @@ class DateHelper
      *
      * @return bool
      */
-    public static function isValidTimeStamp($timestamp)
+    public function isValidTimeStamp($timestamp)
     {
         return ((string) (int) $timestamp === $timestamp)
             && ($timestamp <= PHP_INT_MAX)
