+++ alma/phpunit.ci.xml
@@ -22,6 +22,9 @@
             <exclude>
                 <directory>../alma/vendor</directory>
             </exclude>
+            <exclude>
+                <directory>../alma/tests</directory>
+            </exclude>
         </whitelist>
     </filter>
